refactor: decorations file naming convention
