+++ src/decorations/inline-error.ts
